AS7-1679 Fix auto-timer method handling on beans
